# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/documentos-contratos/contrato/{contratoId}": {
      "get": {
        "operationId": "DocumentoContratoController_findByContratoId",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve los documentos contratos para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentoContrato"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron documentos contratos para el contrato especificado."
          }
        },
        "summary": "Obtener documentos contratos por id de contrato",
        "tags": [
          "documentos-contratos"
        ]
      }
    },
...
    "/contratistas/contrato/{contratoId}": {
      "get": {
        "operationId": "ContratistaController_findByContrato",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el contratista para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contratista"
                }
              }
            }
          },
          "404": {
            "description": "No se encontró contratista para el contrato especificado."
          }
        },
        "summary": "Obtener contratista por id de contrato",
        "tags": [
          "contratistas"
        ]
      }
    },
    "/estados-contrato": {
      "post": {
        "operationId": "EstadoContratoController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEstadoContratoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "El estado de contrato ha sido creado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoContrato"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          }
        },
        "summary": "Crear un nuevo estado de contrato",
        "tags": [
          "estados-contrato"
        ]
      },
      "get": {
        "operationId": "EstadoContratoController_findAll",
        "parameters": [
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "Campos a retornar (separados por comas)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Campo por el cual ordenar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Dirección del ordenamiento",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Número de registros a retornar",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Número de registros a saltar",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Filtros dinámicos en formato JSON",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "description": "Relaciones a Incliur (separadas por comas)",
            "schema": {
              "example": "estados,contratista",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve todos los estados de contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EstadoContrato"
                  }
                }
              }
            }
          }
        },
        "summary": "Obtener todos los estados de contrato",
        "tags": [
          "estados-contrato"
        ]
      }
    },
    "/estados-contrato/{id}": {
      "get": {
        "operationId": "EstadoContratoController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el estado de contrato.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoContrato"
                }
              }
            }
          },
          "404": {
            "description": "Estado de contrato no encontrado."
          }
        },
        "summary": "Obtener un estado de contrato por id",
        "tags": [
          "estados-contrato"
        ]
      },
      "patch": {
        "operationId": "EstadoContratoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEstadoContratoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "El estado de contrato ha sido actualizado exitosamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoContrato"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta."
          },
          "404": {
            "description": "Estado de contrato no encontrado."
          }
        },
        "summary": "Actualizar un estado de contrato",
        "tags": [
          "estados-contrato"
        ]
      },
      "delete": {
        "operationId": "EstadoContratoController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "El estado de contrato ha sido eliminado exitosamente."
          },
          "404": {
            "description": "Estado de contrato no encontrado."
          }
        },
        "summary": "Eliminar un estado de contrato",
        "tags": [
          "estados-contrato"
        ]
      }
    },
    "/estados-contrato/contrato/{contratoId}": {
      "get": {
        "operationId": "EstadoContratoController_findByContratoGeneral",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve los estados de contrato para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EstadoContrato"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron estados de contrato para el contrato especificado."
          }
        },
        "summary": "Obtener estados de contrato por id de contrato",
        "tags": [
          "estados-contrato"
        ]
      }
    },
    "/estados-contrato/contrato/{contratoId}/actual": {
      "get": {
        "operationId": "EstadoContratoController_findCurrentEstado",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el estado actual del contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoContrato"
                }
              }
            }
          },
          "404": {
            "description": "No se encontró estado actual para el contrato especificado."
          }
        },
        "summary": "Obtener el estado actual de un contrato",
        "tags": [
          "estados-contrato"
        ]
      }
    },
...
    "/lugares-ejecucion/contrato/{contratoId}": {
      "get": {
        "operationId": "LugarEjecucionController_findByContratoGeneralId",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve los lugares de ejecución para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LugarEjecucion"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron lugares de ejecución para el contrato especificado."
          }
        },
        "summary": "Obtiene el lugar de ejecución por id de contrato",
        "tags": [
          "lugares-ejecucion"
        ]
      }
    },
...
    "/cdp/contrato/{contratoId}": {
      "get": {
        "operationId": "CdpController_findByContrato",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve los CDPs para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Cdp"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron CDPs para el contrato especificado."
          }
        },
        "summary": "Obtener CDPs por ID de contrato",
        "tags": [
          "cdp"
        ]
      }
    },
...
    "/solicitantes/contrato/{contratoId}": {
      "get": {
        "operationId": "SolicitanteController_findByContratoGeneralId",
        "parameters": [
          {
            "name": "contratoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve los lugares de ejecución para el contrato especificado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SolicitanteEntity"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No se encontraron lugares de ejecución para el contrato especificado."
          }
        },
        "summary": "Obtiene el lugar de ejecución por id de contrato",
        "tags": [
          "solicitantes"
        ]
      }
    }
  },
  "info": {
    "title": "Gestion Contractual CRUD",
    "description": "API para la gestión contractual",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "DocumentoContrato": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Identificador único del documento"
          },
          "tipo_documento_id": {
            "type": "number",
            "example": 1,
            "description": "ID Tipo Documento - Parámetros CRUD"
          },
          "documento_enlace": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Enlace al documento"
          },
          "documento_id": {
            "type": "number",
            "example": 1,
            "description": "Documento ID - Gestor Documental"
          }
        },
        "required": [
          "id",
          "documento_id"
        ]
      },
...
      "Contratista": {
        "type": "object",
        "properties": {}
      },
...
      "CreateEstadoContratoDto": {
        "type": "object",
        "properties": {}
      },
      "EstadoContrato": {
        "type": "object",
        "properties": {}
      },
      "UpdateEstadoContratoDto": {
        "type": "object",
        "properties": {}
      },
...
      "LugarEjecucion": {
        "type": "object",
        "properties": {}
      },
...
      "Cdp": {
        "type": "object",
        "properties": {}
      },
...
      "SolicitanteEntity": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}